disable message delete for each queue

+++ src/main/java/io/relution/jenkins/awssqs/threading/SQSQueueMonitorImpl.java
@@ -204,7 +204,7 @@ public class SQSQueueMonitorImpl implements SQSQueueMonitor {
             return;
         }
 
-        if (this.notifyListeners(messages)) {
+        if (this.notifyListeners(messages) && this.queue.isKeepQueueMessages() == false) {
             this.channel.deleteMessages(messages);
         }
     }
